use reflection to split work
