+++ tests/Unit/Domains/Contact/DAV/Services/ImportVCardTest.php
@@ -6,7 +6,6 @@ use App\Domains\Contact\Dav\Services\ImportVCard;
 use App\Models\User;
 use App\Models\Vault;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
-use Illuminate\Support\Arr;
 use Sabre\VObject\Component\VCard;
 use Sabre\VObject\PHPUnitAssertions;
 use Sabre\VObject\Reader;
@@ -20,7 +19,7 @@ class ImportVCardTest extends TestCase
     /** @test */
     public function it_can_not_import_because_no_firstname_or_nickname_in_vcard()
     {
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
 
         $vcard = new VCard([]);
 
@@ -30,7 +29,7 @@ class ImportVCardTest extends TestCase
     /** @test */
     public function it_can_not_import_because_no_firstname_in_vcard()
     {
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
 
         $vcard = new VCard([
             'N' => ['John', '', '', '', ''],
@@ -42,7 +41,7 @@ class ImportVCardTest extends TestCase
     /** @test */
     public function it_can_not_import_because_empty_firstname_in_vcard()
     {
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
 
         $vcard = new VCard([
             'N' => ';;;;',
@@ -54,7 +53,7 @@ class ImportVCardTest extends TestCase
     /** @test */
     public function it_can_not_import_vcard()
     {
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
 
         $vcard = Reader::read('
 BEGIN:VCARD
@@ -75,7 +74,7 @@ END:VCARD', Reader::OPTION_FORGIVING + Reader::OPTION_IGNORE_INVALID_LINES);
     /** @test */
     public function it_can_not_import_because_empty_nickname_in_vcard()
     {
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
 
         $vcard = new VCard([
             'NICKNAME' => '',
@@ -87,7 +86,7 @@ END:VCARD', Reader::OPTION_FORGIVING + Reader::OPTION_IGNORE_INVALID_LINES);
     /** @test */
     public function it_can_not_import_because_empty_fullname_in_vcard()
     {
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
 
         $vcard = new VCard([
             'FN' => '',
@@ -99,7 +98,7 @@ END:VCARD', Reader::OPTION_FORGIVING + Reader::OPTION_IGNORE_INVALID_LINES);
     /** @test */
     public function it_can_import_firstname()
     {
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
 
         $vcard = new VCard([
             'N' => ['', 'John', '', '', ''],
@@ -111,7 +110,7 @@ END:VCARD', Reader::OPTION_FORGIVING + Reader::OPTION_IGNORE_INVALID_LINES);
     /** @test */
     public function it_can_import_nickname()
     {
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
 
         $vcard = new VCard([
             'NICKNAME' => 'John',
@@ -123,7 +122,7 @@ END:VCARD', Reader::OPTION_FORGIVING + Reader::OPTION_IGNORE_INVALID_LINES);
     /** @test */
     public function it_can_import_fullname()
     {
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
 
         $vcard = new VCard([
             'FN' => 'John Doe',
@@ -135,7 +134,7 @@ END:VCARD', Reader::OPTION_FORGIVING + Reader::OPTION_IGNORE_INVALID_LINES);
     /** @test */
     public function it_returns_an_unknown_name_if_no_name_is_in_entry()
     {
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
 
         $vcard = new VCard([
             'EMAIL' => 'john@',
@@ -150,7 +149,7 @@ END:VCARD', Reader::OPTION_FORGIVING + Reader::OPTION_IGNORE_INVALID_LINES);
     /** @test */
     public function it_returns_a_name_for_N()
     {
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
 
         $vcard = new VCard([
             'N' => ['John', 'Doe', '', '', ''],
@@ -163,7 +162,7 @@ END:VCARD', Reader::OPTION_FORGIVING + Reader::OPTION_IGNORE_INVALID_LINES);
     /** @test */
     public function it_returns_a_name_for_N_incomplete()
     {
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
 
         $vcard = new VCard([
             'N' => ['John', 'Doe'],
@@ -176,7 +175,7 @@ END:VCARD', Reader::OPTION_FORGIVING + Reader::OPTION_IGNORE_INVALID_LINES);
     /** @test */
     public function it_returns_a_name_for_NICKNAME()
     {
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
 
         $vcard = new VCard([
             'NICKNAME' => 'John',
@@ -189,7 +188,7 @@ END:VCARD', Reader::OPTION_FORGIVING + Reader::OPTION_IGNORE_INVALID_LINES);
     /** @test */
     public function it_returns_a_name_for_FN()
     {
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
 
         $vcard = new VCard([
             'FN' => 'John Doe',
@@ -202,7 +201,7 @@ END:VCARD', Reader::OPTION_FORGIVING + Reader::OPTION_IGNORE_INVALID_LINES);
     /** @test */
     public function it_formats_value()
     {
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
 
         $result = $this->invokePrivateMethod($importVCard, 'formatValue', ['']);
         $this->assertNull($result);
@@ -219,7 +218,7 @@ END:VCARD', Reader::OPTION_FORGIVING + Reader::OPTION_IGNORE_INVALID_LINES);
     {
         $author = User::factory()->create();
         $vault = $this->createVaultUser($author, Vault::PERMISSION_EDIT);
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
         $importVCard->accountId = $author->account_id;
         $this->setPrivateValue($importVCard, 'author', $author);
         $this->setPrivateValue($importVCard, 'vault', $vault);
@@ -235,158 +234,11 @@ END:VCARD', Reader::OPTION_FORGIVING + Reader::OPTION_IGNORE_INVALID_LINES);
     }
 
     /** @test */
-    public function it_imports_names_N()
-    {
-        $importVCard = new ImportVCard;
-
-        $vcard = new VCard([
-            'N' => ['Doe', 'John', 'Jane', '', ''],
-        ]);
-        $contact = $this->invokePrivateMethod($importVCard, 'importNames', [[], $vcard]);
-
-        $this->assertEquals('John', $contact['first_name']);
-        $this->assertEquals('Doe', $contact['last_name']);
-        $this->assertEquals('Jane', $contact['middle_name']);
-    }
-
-    /** @test */
-    public function it_imports_names_NICKNAME()
-    {
-        $importVCard = new ImportVCard;
-
-        $vcard = new VCard([
-            'NICKNAME' => 'John',
-        ]);
-        $contact = $this->invokePrivateMethod($importVCard, 'importNames', [[], $vcard]);
-
-        $this->assertEquals('John', $contact['first_name']);
-    }
-
-    /** @test */
-    public function it_imports_names_FN()
-    {
-        $author = User::factory()->create();
-        $importVCard = new ImportVCard;
-        $this->setPrivateValue($importVCard, 'author', $author);
-
-        $vcard = new VCard([
-            'FN' => 'John Doe',
-        ]);
-        $contact = $this->invokePrivateMethod($importVCard, 'importNames', [[], $vcard]);
-
-        $this->assertEquals('John', $contact['first_name']);
-        $this->assertEquals('Doe', $contact['last_name']);
-    }
-
-    /** @test */
-    public function it_imports_names_FN_last()
-    {
-        $author = User::factory()->create([
-            'name_order' => '%last_name% %first_name%',
-        ]);
-        $importVCard = new ImportVCard;
-        $this->setPrivateValue($importVCard, 'author', $author);
-
-        $vcard = new VCard([
-            'FN' => 'John Doe',
-        ]);
-        $contact = $this->invokePrivateMethod($importVCard, 'importNames', [[], $vcard]);
-
-        $this->assertEquals('Doe', $contact['first_name']);
-        $this->assertEquals('John', $contact['last_name']);
-    }
-
-    /** @test */
-    public function it_imports_names_FN_extra_space()
-    {
-        $author = User::factory()->create();
-        $importVCard = new ImportVCard;
-        $this->setPrivateValue($importVCard, 'author', $author);
-
-        $vcard = new VCard([
-            'FN' => 'John  Doe',
-        ]);
-        $contact = $this->invokePrivateMethod($importVCard, 'importNames', [[], $vcard]);
-
-        $this->assertEquals('John', $contact['first_name']);
-        $this->assertEquals('Doe', $contact['last_name']);
-    }
-
-    /** @test */
-    public function it_imports_name_FN()
-    {
-        $author = User::factory()->create();
-        $importVCard = new ImportVCard;
-        $this->setPrivateValue($importVCard, 'author', $author);
-
-        $vcard = new VCard([
-            'FN' => 'John',
-            'N' => 'Mike;;;;',
-        ]);
-        $contact = $this->invokePrivateMethod($importVCard, 'importNames', [[], $vcard]);
-
-        $this->assertEquals('John', $contact['first_name']);
-        $this->assertEquals('', Arr::get($contact, 'last_name'));
-    }
-
-    /** @test */
-    public function it_imports_name_FN_last()
-    {
-        $author = User::factory()->create([
-            'name_order' => '%last_name% %first_name%',
-        ]);
-        $importVCard = new ImportVCard;
-        $this->setPrivateValue($importVCard, 'author', $author);
-
-        $vcard = new VCard([
-            'FN' => 'John',
-            'N' => 'Mike;;;;',
-        ]);
-        $contact = $this->invokePrivateMethod($importVCard, 'importNames', [[], $vcard]);
-
-        $this->assertEquals('John', $contact['last_name']);
-        $this->assertEquals('', Arr::get($contact, 'first_name'));
-    }
-
-    /** @test */
-    public function it_imports_names_FN_multiple()
-    {
-        $author = User::factory()->create();
-        $importVCard = new ImportVCard;
-        $this->setPrivateValue($importVCard, 'author', $author);
-
-        $vcard = new VCard([
-            'FN' => 'John Doe Marco',
-            'N' => 'Mike;;;;',
-        ]);
-        $contact = $this->invokePrivateMethod($importVCard, 'importNames', [[], $vcard]);
-
-        $this->assertEquals('John', $contact['first_name']);
-        $this->assertEquals('Doe Marco', $contact['last_name']);
-    }
-
-    /** @test */
-    public function it_imports_uuid_default()
-    {
-        $importVCard = new ImportVCard;
-
-        $vcard = new VCard([
-            'UID' => '31fdc242-c974-436e-98de-6b21624d6e34',
-        ]);
-
-        $contact = [];
-
-        $contact = $this->invokePrivateMethod($importVCard, 'importUid', [$contact, $vcard]);
-
-        $this->assertEquals('31fdc242-c974-436e-98de-6b21624d6e34', $contact['uuid']);
-    }
-
-    /** @test */
     public function it_imports_uuid_contact()
     {
         $author = User::factory()->create();
         $vault = $this->createVaultUser($author, Vault::PERMISSION_EDIT);
-        $importVCard = new ImportVCard;
+        $importVCard = new ImportVCard($this->app);
         $importVCard->accountId = $author->account_id;
         $this->setPrivateValue($importVCard, 'author', $author);
         $this->setPrivateValue($importVCard, 'vault', $vault);
